Extract CmarkVisitor & refactor renderers
